refactor: consolidate optimize/ package into optim/
Move PlaceRouteOptimizer from optimize/place_route.py to
optim/place_route.py, unify the two OptimizationResult dataclasses
into a single canonical type in optim/workflow.py with optional
routing-specific fields, and delete the optimize/ package entirely.
